--- web/src/store/dispatch.ts
@@ -1,241 +1,0 @@
-import {saveAs} from 'file-saver';
-import {push} from 'connected-react-router';
-
-import client, {EvalEventKind, instantiateStreaming,} from '~/services/api';
-
-import {getImportObject, goRun} from '~/services/go';
-import {getWorkerInstance} from "~/services/gorepl";
-import config, {RunTargetConfig, TargetType} from '~/services/config';
-import {DEMO_CODE} from '~/components/editor/props';
-import {isDarkModeEnabled} from "~/utils/theme";
-
-import {PanelState, SettingsState} from './state';
-import {DispatchFn, StateProvider} from "./helpers";
-import {newAddNotificationAction, NotificationType} from "./notifications";
-
-import {
-  Action,
-  ActionType,
-  MonacoParamsChanges,
-  newBuildResultAction,
-  newErrorAction,
-  newImportFileAction,
-  newLoadingAction,
-  newMonacoParamsChangeAction,
-  newPanelStateChangeAction,
-  newProgramWriteAction,
-  newRunTargetChangeAction,
-  newSettingsChangeAction,
-  newToggleThemeAction,
-  newUIStateChangeAction
-} from './actions';
-
-export type Dispatcher = (dispatch: DispatchFn, getState: StateProvider) => void
-
-export type StateDispatch = <V=any,T=string>(v: Action<T, V> | Dispatcher) => void;
-
-/////////////////////////////
-//      Dispatchers        //
-/////////////////////////////
-
-export function newImportFileDispatcher(f: File): Dispatcher {
-  return (dispatch: DispatchFn, _: StateProvider) => {
-    const reader = new FileReader();
-    reader.onload = e => {
-      const data = e.target?.result as string;
-      dispatch(newImportFileAction(f.name, data));
-    };
-
-    reader.onerror = e => {
-      // TODO: replace with a nice modal
-      alert(`Failed to import a file: ${e}`)
-    };
-
-    reader.readAsText(f, 'UTF-8');
-  };
-}
-
-export function newCodeImportDispatcher(name: string, contents: string): Dispatcher {
-  return (dispatch: DispatchFn, _: StateProvider) => {
-    dispatch(newImportFileAction(`${encodeURIComponent(name)}.go`, contents));
-  };
-}
-
-export function newMonacoParamsChangeDispatcher(changes: MonacoParamsChanges): Dispatcher {
-  return (dispatch: DispatchFn, _: StateProvider) => {
-    const current = config.monacoSettings;
-    config.monacoSettings = Object.assign(current, changes);
-    dispatch(newMonacoParamsChangeAction(changes));
-  };
-}
-
-export const newSettingsChangeDispatcher = (changes: Partial<SettingsState>): Dispatcher => (
-  (dispatch: DispatchFn, _: StateProvider) => {
-    if ('useSystemTheme' in changes) {
-      config.useSystemTheme = !!changes.useSystemTheme;
-      changes.darkMode = isDarkModeEnabled();
-    }
-
-    if ('darkMode' in changes) {
-      config.darkThemeEnabled = !!changes.darkMode;
-    }
-
-    if ('enableVimMode' in changes) {
-      config.enableVimMode = !!changes.enableVimMode;
-    }
-
-    dispatch(newSettingsChangeAction(changes));
-  }
-);
-
-export const newRunTargetChangeDispatcher = (cfg: RunTargetConfig): Dispatcher => (
-  (dispatch: DispatchFn) => {
-    config.runTargetConfig = cfg;
-    dispatch(newRunTargetChangeAction(cfg));
-  }
-);
-
-export function newSnippetLoadDispatcher(snippetID?: string): Dispatcher {
-  return async (dispatch: DispatchFn, _: StateProvider) => {
-    if (!snippetID) {
-      dispatch(newImportFileAction('prog.go', DEMO_CODE));
-      return;
-    }
-
-    dispatch(newLoadingAction());
-    try {
-      console.log('loading snippet %s', snippetID);
-      const resp = await client.getSnippet(snippetID);
-      const { fileName, code } = resp;
-      dispatch(newImportFileAction(fileName, code));
-    } catch (err: any) {
-      dispatch(newErrorAction(err.message));
-    }
-  }
-}
-
-export const shareSnippetDispatcher: Dispatcher =
-  async (dispatch: DispatchFn, getState: StateProvider) => {
-    dispatch(newLoadingAction());
-    try {
-      const { code } = getState().editor;
-      const { snippetID } = await client.shareSnippet(code);
-      dispatch(push(`/snippet/${snippetID}`));
-      dispatch(newUIStateChangeAction({ shareCreated: true, snippetId: snippetID }));
-    } catch (err: any) {
-      dispatch(newErrorAction(err.message));
-    }
-  };
-
-export const saveFileDispatcher: Dispatcher =
-  (dispatch: DispatchFn, getState: StateProvider) => {
-    try {
-      const { fileName, code } = getState().editor;
-      const blob = new Blob([code], {
-        type: 'text/plain;charset=utf-8'
-      });
-      saveAs(blob, fileName);
-    } catch (err: any) {
-      dispatch(newAddNotificationAction({
-        id: 'SAVE_FILE_ERROR',
-        type: NotificationType.Error,
-        title: 'Failed to save file',
-        description: err.toString(),
-        canDismiss: true
-      }));
-    }
-  };
-
-export const runFileDispatcher: Dispatcher =
-  async (dispatch: DispatchFn, getState: StateProvider) => {
-    dispatch(newLoadingAction());
-    try {
-      const { settings, editor, runTarget: {target, backend} } = getState();
-      switch (target) {
-        case TargetType.Server:
-          const res = await client.evaluateCode(editor.code, settings.autoFormat, backend);
-          dispatch(newBuildResultAction(res));
-          break;
-        case TargetType.WebAssembly:
-          let resp = await client.build(editor.code, settings.autoFormat);
-          let wasmFile = await client.getArtifact(resp.fileName);
-          let instance = await instantiateStreaming(wasmFile, getImportObject());
-          dispatch({ type: ActionType.EVAL_START });
-          dispatch(newBuildResultAction({ formatted: resp.formatted, events: [] }));
-          goRun(instance)
-            .then(result => console.log('exit code: %d', result))
-            .catch(err => {
-              dispatch(newAddNotificationAction({
-                id: 'WASM_APP_ERROR',
-                type: NotificationType.Error,
-                title: 'Failed to run WebAssembly program',
-                description: err.toString(),
-                canDismiss: true,
-              }));
-            })
-            .finally(() => dispatch({ type: ActionType.EVAL_FINISH }));
-          break;
-        case TargetType.Interpreter:
-          try {
-            const worker = await getWorkerInstance(dispatch, getState);
-            await worker.runProgram(editor.code);
-          } catch (err: any) {
-            dispatch(newErrorAction(err.message ?? err.toString()));
-          }
-
-          break;
-        default:
-          dispatch(newErrorAction(`AppError: Unknown Go runtime type "${target}"`));
-      }
-    } catch (err: any) {
-      dispatch(newErrorAction(err.message));
-    }
-  };
-
-export const formatFileDispatcher: Dispatcher =
-  async (dispatch: DispatchFn, getState: StateProvider) => {
-    dispatch(newLoadingAction());
-    try {
-      // Format code using GoTip is enabled to support
-      // any syntax changes from unstable Go specs.
-      const { editor: {code}, runTarget: { backend } } = getState();
-      const res = await client.formatCode(code, backend);
-
-      if (res.formatted) {
-        dispatch(newBuildResultAction(res));
-      }
-    } catch (err: any) {
-      dispatch(newErrorAction(err.message));
-    }
-  };
-
-export const dispatchToggleTheme: Dispatcher =
-  (dispatch: DispatchFn, getState: StateProvider) => {
-    const { darkMode } = getState().settings;
-    config.darkThemeEnabled = !darkMode;
-    dispatch(newToggleThemeAction())
-  };
-
-export const dispatchPanelLayoutChange = (changes: Partial<PanelState>): Dispatcher => (
-  (dispatch: DispatchFn, getState: StateProvider) => {
-    const { panel } = getState();
-    config.panelLayout = { ...panel, ...changes };
-    dispatch(newPanelStateChangeAction(changes));
-  }
-);
-
-//////////////////////////////////
-//          Adapters            //
-//////////////////////////////////
-
-export const createGoConsoleAdapter = (dispatch: DispatchFn) =>
-({
-  log: (eventType: EvalEventKind, message: string) => {
-    console.log('%s:\t%s', eventType, message);
-    dispatch(newProgramWriteAction({
-      Kind: eventType,
-      Message: message,
-      Delay: 0,
-    }));
-  }
-});
