+++ web/src/store/index.ts
@@ -1,5 +1,5 @@
 export * from './actions';
 export * from './state';
-export * from './dispatch';
+export * from './dispatchers';
 export type { DispatchFn } from './helpers';
 export { configureStore } from './configure';
